src/main/java/com/soave/backend_performance_challenge/controller/UserController.java
package com.soave.backend_performance_challenge.controller;

import com.soave.backend_performance_challenge.model.dto.UserDto;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.net.URI;
import java.util.UUID;

@RestController
@RequestMapping("/users")
public class UserController {

    @PostMapping
    public ResponseEntity<UserDto> createUser(@RequestBody @Valid UserDto userDto) {
        URI location = URI.create("/pessoas/" + userDto.id());
        return ResponseEntity.created(location).body(userDto);
    }
}
